dept delete

+++ pkg/services/dept.go
@@ -2,17 +2,20 @@ package services
 
 import (
 	"errors"
+	"fmt"
 	"meal-management/pkg/domain"
 	"meal-management/pkg/models"
 )
 
 type DeptService struct {
-	repo domain.IDeptRepo
+	repo  domain.IDeptRepo
+	repo2 domain.IEmployeeRepo
 }
 
-func DeptServiceInstance(deptRepo domain.IDeptRepo) domain.IDeptService {
+func DeptServiceInstance(deptRepo domain.IDeptRepo, employeeRepo domain.IEmployeeRepo) domain.IDeptService {
 	return &DeptService{
-		repo: deptRepo,
+		repo:  deptRepo,
+		repo2: employeeRepo,
 	}
 }
 
@@ -30,8 +33,15 @@ func (service *DeptService) UpdateDepartment(dept *models.Department) error {
 }
 
 func (service *DeptService) DeleteDepartment(deptID int) error {
+	employee, err := service.repo2.GetEmployeeByDepartment(deptID)
+	if err != nil {
+		return err
+	}
+	if employee != nil {
+		return fmt.Errorf("department still has %d employees", len(employee))
+	}
 	if err := service.repo.DeleteDepartment(deptID); err != nil {
-		return errors.New("department was not created")
+		return errors.New("department does not exist")
 	}
 	return nil
 }
